Moved encryption functions to seperate file
